requirements.txt
absl-py==2.4.0
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.4
colorlog==4.7.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.76.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lightgbm==4.6.0
Markdown==3.10.1
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.16.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==26.0
pandas==3.0.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.2
protobuf==6.33.5
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
ray==2.6.3
recbole==1.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
sentry-sdk==2.52.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.24.2
Werkzeug==3.1.5
xgboost==3.1.3
